Move thermal_reponse_factors to heat_transfer module
